+++ src/main/java/com/ArthurGrand/module/notification/events/NotificationEvent.java
@@ -10,7 +10,7 @@ import java.time.Instant;
 @Data
 @AllArgsConstructor
 public class NotificationEvent {
-    private final Long tenantId;
+    private final int tenantId;
     private final String email;
     private final String subject;
     private final String message;
