+++ src/main/java/com/ArthurGrand/module/notification/serviceImp/AppNotificationServiceImp.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Service;
 @Service
 public class AppNotificationServiceImp implements AppNotificationService {
     @Override
-    public void sendAppNotification(Long tenantId, String message) {
+    public void sendAppNotification(int tenantId, String message) {
         // Log or save to DB
         System.out.println("APP NOTIFICATION: [" + tenantId + "] " + message);
     }
